Patcher to check if an item is expired before processing (#589)
Patcher to check if an item is expired before processing

This is not a perfect change in that the patcher keeps running
for a quite long time and we only check at the beginnign. However,
it will ensure that we won't endlessly retry the same failing task
over and over and just burn LLM-credits.