More robust copying of live corpus (#686)
* More robust copying of live corpus

As the corpus might change due to fuzzers modifying it, the merger bot
can occasionally fail due to missing local files, or files that where
copied from remote that no longer exists in local. For the merger to
operate ideally, we want be sure that the remote files are all included
in the merge, while local files we just take what we get. There will
be a next attempt soon.

* Sample the local files to merge to remote corpus (#691)

This is done to try and limit the time the merge takes. This will allow
for more parallel mergers making progress and not stalling each other on
the same harness.