Initial orchestrator deduplication (#838)
* Initial orchestrator deduplication

This is the first phase handling of PoV duplicates. PoVs having
identical crashes are stopped at the fuzzer. Others migrate to the
orchestrator. In the orchestrator PoVs are compared to see if they are
'similar' to any other PoV. If so they are bundled and each are used to
validate patch quality before submitting.

This PR basically acts as a filter for only submitting a patch for any
PoV considered similar before the patch can be submitted the first time.

* Make sure duplicates are only considered for the same task

---------

Co-authored-by: Michael D Brown <michael.brown@trailofbits.com>