patcher: do not rely on Agent's challenge (#841)
The patcher now keeps track of the built challenges in PatchAttempt,
thus it is not necessary anymore to use self.challenge everywhere. We
keep using it as a read-only challenge to extract read-only info like
project name, task_id, etc.

The context retriever creates a copy as needed and operates on that,
while the other agents either just use the read-only task or use the
PatchAttempt.built_challenges.