common: add first frame line number in the crash set (#152)
This will differentiate between vulnerabilities in different lines of
the same function, so that they all go to the tracer-bot for validation.
In this way, if there is a DELTA mode challenge and one vuln is
present both before and after the delta, but another one is introduced
by the delta, we are able to validate them separately and actually find
the vulnerability that is introduced by the delta (and not discard it in
the fuzzer-bot due to being already present in the CrashSet).

It means the tracer-bot will receive slightly more crashes to validate,
but still a reasonable number.

Co-authored-by: Michael D Brown <michael.brown@trailofbits.com>