fuzzer_runner: split fuzzer's dependencies
Before this commit, the fuzzer dependend on clusterfuzz, which uses
protobuf 3.20. Since fuzzer-bot depended on common subpackage as well
and also common (and the other packages) require protobuf, we had to use
protobuf 3.20 everywhere. This old dependency however means that a lot
of packages can't be used in their "newer" versions, because they depend
on newer protobuf versions.

This commit splits the fuzzer into a separate fuzzer-runner that is
executed in a separate process (sidecar container) and communicate with
the rest of the system through REST API. fuzzer-runner executed the
actual clusterfuzz-heavy operations (e.g. fuzzing) and isolates the
clusterfuzz dependency, so that the rest of the system can use newer
protobuf version.

* split fuzzer-bot in fuzzer-runner and fuzzer-bot
* have a "full" optional dependency group in common, including openlit
  and protobuf, so that fuzzer-runner can use the lite version of common
  without bring those heavy dep
* move FuzzConfiguration/BuildConfiguration in a separate common file
  that doesn't require to load all protobuf files. Again, in this way
  other components can just depend on the common "lite" version and not
  require protobuf stuff
* add fuzzer-runner as a sidecar container to fuzzer-bot pod, so
  communication is very fast even though it happens over REST API
* add RunnerProxy class in fuzzer package to provide an interface to
  interact with the fuzzer-runner over REST API